restructure app and promote v2 to main
