+++ config.py
@@ -1,0 +1,12 @@
+TIME_INCREMENT = 30  # in_minutes
+NUMBER_OF_TIMES_TO_ADVANCE_WORLD = 10
+
+# Used for debugging
+ENABLE_PROMPT_CONTEXT_LOGGING = False
+ENABLE_PROMPT_LOGGING = False
+ENABLE_METHOD_CALLED_LOGGING = False
+ENABLE_OPENAI_RESPONSE_LOGGING = False
+
+# Generally should leave enabled
+ENABLE_WORLD_EVENT_LOGGING = True
+ENABLE_AGENT_EVENT_LOGGING = True
