+++ memory.py
