+++ pinecone_manager.py
