+++ readme.md
@@ -36,50 +36,19 @@ Edit the `.env` file to put in your OpenAI private key
 
 ## Configuration
 
-1. Create an OpenAI account and obtain your API key.
+Edit your config variables in `config.py`:
 
-2. Replace the placeholder API key in `main.py` with your own API key.
-```python
-openai.api_key = "your-api-key"
-```
+- Set `TIME_INTERVAL` to however many minutes of in-game time should pass per step
+- Set `NUMBER_OF_TIMES_TO_ADVANCE_WORLD`
 
-## Usage
+## Usage / Run
 Run the main script.
 
 `python main.py`
 
-Follow the on-screen instructions to interact with the agents. You can issue the following commands:
 
-- talk: Make the agents have a conversation.
-- recall: Display an agent's long-term memories.
-- add_task: Add a task for an agent to perform.
-- perform_tasks: (Deprecated) Tasks are now performed automatically during the step command.
-- step: Move the simulation forward by one time unit.
-- quit: Exit the simulation.
-
-## Files
-- main.py: The main script that runs the simulation.
-- agent.py: Contains the Agent class definition, which represents an individual agent.
-- scheduler.py: Contains the Scheduler class definition, which manages the agents' actions over time.
-
-## Roadmap
-
-- [x] Agent
-- [x] Global time
-- [x] Dialogue
-- [x] Task management
-- [x] Task extraction from dialogue
-- [x] more realistic dialogue
-- [x] OpenAI setup + dialogue
-- [x] Simple memory system (long term + short term)
-- [ ] Daily planning
-- [ ] Goal setting, objectives, and long term planning
-- [ ] Robust long term and short term memory truncating
-- [ ] God mode
-- [ ] Environment interaction
-- [ ] More personality, traits, and emotions
-- [ ] Suggestions?
+## Debugging
+To have more robust debugging, you can turn on various levels of logging using the variables found in the `config.py`, which determines system logging in the `handle_logging` method in `helpers.py`.
 
 ## License
 This project is licensed under the MIT License.
-
