--- v2/config.py
@@ -1,6 +1,0 @@
-ENABLE_PROMPT_CONTEXT_LOGGING = True
-ENABLE_PROMPT_LOGGING = True
-ENABLE_METHOD_CALLED_LOGGING = True
-ENABLE_OPENAI_RESPONSE_LOGGING = True
-ENABLE_WORLD_EVENT_LOGGING = True
-ENABLE_AGENT_EVENT_LOGGING = True
